Remove old src/thread_local
