+++ examples/thread_local.rs
@@ -2,8 +2,16 @@ use std::sync::mpsc::channel;
 use std::sync::Arc;
 use std::thread;
 
+use mcslock::raw::spins::Mutex;
+
 // Requires that the `thread_local` feature is enabled.
-use mcslock::thread_local::spins::Mutex;
+mcslock::thread_local_node! {
+    // * Allows multiple static definitions, must be separated with semicolons.
+    // * Visibility is optional (private by default).
+    // * Requires `static` keyword and a UPPER_SNAKE_CASE name.
+    pub static NODE;
+    static UNUSED_NODE;
+}
 
 const N: usize = 10;
 
@@ -27,7 +35,7 @@ fn main() {
             // threads to ever fail while holding the lock.
             //
             // Data is exclusively accessed by the guard argument.
-            data.lock_with(|mut data| {
+            data.lock_with_local(&NODE, |mut data| {
                 *data += 1;
                 if *data == N {
                     tx.send(()).unwrap();
@@ -38,6 +46,6 @@ fn main() {
     }
     let _message = rx.recv().unwrap();
 
-    let count = data.lock_with(|guard| *guard);
+    let count = data.lock_with_local(&NODE, |guard| *guard);
     assert_eq!(count, N);
 }
